+++ src-ui/app/services/telemetry.service.ts
@@ -1,27 +1,20 @@
-import { Injectable, isDevMode } from '@angular/core';
-import { TelemetryManifest } from '../models/telemetry-manifest';
+import { Injectable } from '@angular/core';
 import { Store } from 'tauri-plugin-store-api';
 import { SETTINGS_FILE, SETTINGS_KEY_TELEMETRY_SETTINGS } from '../globals';
 import {
   asyncScheduler,
   BehaviorSubject,
-  debounceTime,
   distinctUntilChanged,
-  filter,
-  firstValueFrom,
   map,
   Observable,
-  pairwise,
   switchMap,
   throttleTime,
 } from 'rxjs';
 import { TELEMETRY_SETTINGS_DEFAULT, TelemetrySettings } from '../models/telemetry-settings';
 import { migrateTelemetrySettings } from '../migrations/telemetry-settings.migrations';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
+import { HttpClient } from '@angular/common/http';
 import { cloneDeep } from 'lodash';
 import { AppSettingsService } from './app-settings.service';
-import { getVersion } from '../utils/app-utils';
-import { debug, info } from 'tauri-plugin-log-api';
 import { invoke } from '@tauri-apps/api';
 import { trackEvent } from '@aptabase/tauri';
 
@@ -34,8 +27,6 @@ export class TelemetryService {
     TELEMETRY_SETTINGS_DEFAULT
   );
   public settings: Observable<TelemetrySettings> = this._settings.asObservable();
-  private manifest?: TelemetryManifest;
-  private timeout?: NodeJS.Timeout;
 
   constructor(private http: HttpClient, private appSettings: AppSettingsService) {}
 
@@ -51,20 +42,6 @@ export class TelemetryService {
         })
       )
       .subscribe();
-    if (!isDevMode()) {
-      this.timeout = setTimeout(() => this.scheduleTelemetry(), 10000);
-      // Send heartbeat when language setting has changed
-      this.appSettings.settings
-        .pipe(
-          map((settings) => settings.userLanguage),
-          pairwise(),
-          filter(([oldLang, newLang]) => oldLang !== newLang),
-          debounceTime(20000)
-        )
-        .subscribe(() => this.scheduleTelemetry());
-    } else {
-      debug('[Telemetry] Disabling telemetry in dev mode');
-    }
 
     addEventListener('unhandledrejection', (e) => {
       this.trackEvent('ui_promise_rejected', {
@@ -97,88 +74,6 @@ export class TelemetryService {
     const newSettings = Object.assign(cloneDeep(this._settings.value), settings);
     this._settings.next(newSettings);
     await this.saveSettings();
-  }
-
-  async scheduleTelemetry() {
-    if (this.timeout) clearTimeout(this.timeout);
-    if (await this.sendTelemetry()) {
-      // If successful, schedule next heartbeat in 24 hours
-      this.timeout = setTimeout(() => this.scheduleTelemetry(), 1000 * 60 * 60 * 24);
-    } else {
-      // If unsuccessful, schedule next heartbeat in 30 minutes
-      this.timeout = setTimeout(() => this.scheduleTelemetry(), 1000 * 60 * 30);
-    }
-  }
-
-  async sendTelemetry(): Promise<boolean> {
-    try {
-      const settings = this._settings.value;
-      // Stop if telemetry is not enabled
-      if (!settings.enabled) return false;
-      // Fetch manifest if needed
-      if (!this.manifest) {
-        await this.fetchManifest();
-        if (!this.manifest) return false;
-      }
-      // Determine version and language to send
-      const version = await getVersion();
-      const lang = await firstValueFrom(this.appSettings.settings).then(
-        (settings) => settings.userLanguage
-      );
-      // Stop if last heartbeat was sent less than 24 hours ago for the same version and language
-      if (
-        settings.lastVersion === version &&
-        settings.lastLang === lang &&
-        Date.now() - this._settings.value.lastHeartbeat < 1000 * 60 * 60 * 24
-      )
-        return false;
-      // Send heartbeat
-      let headers = new HttpHeaders();
-      Object.entries(this.manifest.v1.heartbeatHeaders).forEach(
-        ([key, value]) => (headers = headers.set(key, value))
-      );
-      info(
-        '[Telemetry] Sending heartbeat (id=' +
-          settings.telemetryId +
-          ',version=' +
-          version +
-          ', lang=' +
-          lang +
-          ')'
-      );
-      const response = await firstValueFrom(
-        this.http.post<{ status: string }>(
-          this.manifest.v1.heartbeatUrl,
-          {
-            telemetryId: settings.telemetryId,
-            version,
-            lang,
-          },
-          { headers, observe: 'response' }
-        )
-      );
-      if (response.status !== 200 || response.body?.status !== 'ok') return false;
-      // Update last heartbeat
-      this._settings.next({
-        ...settings,
-        lastHeartbeat: Date.now(),
-        lastVersion: version,
-        lastLang: lang,
-      });
-      await this.saveSettings();
-      return true;
-    } catch (e) {
-      return false;
-    }
-  }
-
-  async fetchManifest() {
-    info('[Telemetry] Fetching manifest');
-    this.manifest = await firstValueFrom(
-      this.http.get<TelemetryManifest>(
-        'https://gist.githubusercontent.com/Raphiiko/675fcd03e1e22c2514951ef21c99e4d5/raw/1ba8bc77ab994da8f86f5cf184dd9c79e77f481d/oyasumi_telemetry_manifest.json'
-      )
-    );
   }
 
   async trackEvent(event: string, props: { [key: string]: string | number }) {