+++ src-ui/build.ts
@@ -5,4 +5,4 @@
 
 export type BuildFlavour = 'DEV' | 'STANDALONE' | 'STEAM' | 'STEAM_CN';
 export const FLAVOUR: BuildFlavour = 'DEV';
-export const BUILD_ID = 'devbuild';
+export const BUILD_ID = 'fda0796';